requirements.txt
fpdf2 == 2.7.6
nltk == 3.8.1 